Create working svelte select component
