+++ package.json
@@ -57,6 +57,7 @@
 		"pdf.js-extract": "^0.2.1",
 		"pocketbase": "^0.21.1",
 		"svelte-radix": "^1.1.0",
+		"svelte-select": "^5.8.3",
 		"svelte-sonner": "^0.3.24",
 		"tailwind-merge": "^2.3.0",
 		"tailwind-variants": "^0.2.1",
